Enum network::ConsensusFork

pub enum ConsensusFork {
    BitcoinCore,
    BitcoinCash(BitcoinCashConsensusParams),
}

Concurrent consensus rule forks.

Variants
BitcoinCore

No fork.

BitcoinCash(BitcoinCashConsensusParams)

Bitcoin Cash (aka UAHF). u32 is height of the first block, for which new consensus rules are applied. Briefly: no SegWit + blocks up to 8MB + replay protection. Technical specification: UAHF Technical Specification - https://github.com/Bitcoin-UAHF/spec/blob/master/uahf-technical-spec.md BUIP-HF Digest for replay protected signature verification across hard forks - https://github.com/Bitcoin-UAHF/spec/blob/master/replay-protected-sighash.md

Methods

impl ConsensusFork

pub fn absolute_maximum_block_size() -> usize

Absolute (across all forks) maximum block size. Currently is 8MB for post-HF BitcoinCash

pub fn absolute_maximum_block_sigops() -> usize

Absolute (across all forks) maximum number of sigops in single block. Currently is max(sigops) for 8MB post-HF BitcoinCash block

pub fn witness_scale_factor() -> usize

Witness scale factor (equal among all forks)

pub fn activation_height(&self) -> u32

pub fn max_transaction_size(&self) -> usize

pub fn min_block_size(&self, height: u32) -> usize

pub fn max_block_size(&self, height: u32, median_time_past: u32) -> usize

pub fn max_block_sigops(&self, height: u32, block_size: usize) -> usize

pub fn max_block_sigops_cost(&self, height: u32, block_size: usize) -> usize

pub fn max_block_weight(&self, _height: u32) -> usize