Enum network::Network

pub enum Network {
    Mainnet,
    Testnet,
    Regtest,
    Unitest,
    Other(u32),
}

Bitcoin network

Variants
Mainnet

The original and main network for Bitcoin transactions, where satoshis have real economic value.

Testnet

The main bitcoin testnet.

Regtest

Bitcoin regtest network.

Unitest

Testnet for unittests, proof of work difficulty is almost 0

Other(u32)

Any other network. By default behaves like bitcoin mainnet.

Methods

impl Network

pub fn magic(&self, fork: &ConsensusFork) -> Magic

pub fn max_bits(&self) -> U256

pub fn port(&self) -> u16

pub fn rpc_port(&self) -> u16

pub fn genesis_block(&self) -> Block

pub fn default_verification_edge(&self) -> H256