Module network::fuzzing

Functions

exchange_handshake_input
Builds two HandshakeMsg and serializes the second one. It is the input expected for the fuzzer.

fuzz_network_handshake_protocol_exchange
Fuzzing the handshake protocol, which negotiates protocols supported by both the client and the server. At the moment, fuzzing the client or the server leads to the same logic.

fuzz_network_handshake_protocol_negotiation
Same function as fuzz_network_handshake_protocol_exchange except that the network exchange is skipped, letting us skip BCS deserialization (and potentially other logic) and fuzz the negotiation of protocols directly.

perform_handshake_input
Builds two HandshakeMsg and serializes the second one. It is the input expected for the fuzzer.