Module network::interface

Internal Rust Network API for Consensus and Mempool

Structs

ConsensusNetworkEvents
The interface from Network to Consensus layer.

ConsensusNetworkSender
The interface from Consensus to Networking layer.

MempoolNetworkEvents
The interface from Network to Mempool layer.

MempoolNetworkSender
The interface from Mempool to Networking layer.

StateSynchronizerNetworkEvents
The interface from Network to StateSynchronizer.

StateSynchronizerNetworkSender
The interface from StateSynchronizer to Networking layer.

Enums

Event
A wrapper around the various types of messages that the Networking layer will be sending to either Consensus or Mempool.

Constants

CONSENSUS_DIRECT_SEND_PROTOCOL
Protocol id for consensus direct-send calls

CONSENSUS_RPC_PROTOCOL
Protocol id for consensus RPC calls

MEMPOOL_DIRECT_SEND_PROTOCOL
Protocol id for consensus direct-send calls

STATE_SYNCHRONIZER_RPC_PROTOCOL
Protocol id for state synchronizer RPC calls