Module network::logging

Constants

CONNECTION_NOTIFICATION

CONNECTIVITY_MANAGER_LOOP
This file contains constants used for structured logging data types so that there is consistency among structured logs.

CONNECTIVITY_REQUEST

EVENT_ID

NETWORK_CONTEXT
Specific fields for logging

ONCHAIN_DISCOVERY_LOOP

REMOTE_PEER

START

TERMINATION

TYPE
Common terms