Enum network::noise::error::NoiseHandshakeError

pub enum NoiseHandshakeError {

    MissingServerPublicKey,
    BuildClientHandshakeMessageFailed(NoiseError),
    ClientWriteFailed(Error),
    ClientReadFailed(Error),
    ClientFlushFailed(Error),
    ClientFinalizeFailed(NoiseError),
    ServerReadFailed(Error),
    InvalidClientPeerId(String),
    SelfDialDetected,
    ClientExpectingDifferentPubkey(ShortHexStr, String),
    ServerParseClient(ShortHexStr, NoiseError),
    UnauthenticatedClientPubkey(ShortHexStr, String),
    UnauthenticatedClient(ShortHexStr, PeerId),
    ClientPeerIdMismatch(ShortHexStr, PeerId, PeerId),
    MissingAntiReplayTimestamp(ShortHexStr),
    ServerReplayDetected(ShortHexStr, u64),
    BuildServerHandshakeMessageFailed(ShortHexStr, NoiseError),
    ServerWriteFailed(ShortHexStr, Error),

}

Different errors than can be raised when negotiating a Noise handshake.

Variants

MissingServerPublicKey

BuildClientHandshakeMessageFailed(NoiseError)

Tuple Fields of BuildClientHandshakeMessageFailed

0: NoiseError

ClientWriteFailed(Error)

Tuple Fields of ClientWriteFailed

0: Error

ClientReadFailed(Error)

Tuple Fields of ClientReadFailed

0: Error

ClientFlushFailed(Error)

Tuple Fields of ClientFlushFailed

0: Error

ClientFinalizeFailed(NoiseError)

Tuple Fields of ClientFinalizeFailed

0: NoiseError

ServerReadFailed(Error)

Tuple Fields of ServerReadFailed

0: Error

InvalidClientPeerId(String)

Tuple Fields of InvalidClientPeerId

0: String

SelfDialDetected

ClientExpectingDifferentPubkey(ShortHexStr, String)

Tuple Fields of ClientExpectingDifferentPubkey

0: ShortHexStr1: String

ServerParseClient(ShortHexStr, NoiseError)

Tuple Fields of ServerParseClient

0: ShortHexStr1: NoiseError

UnauthenticatedClientPubkey(ShortHexStr, String)

Tuple Fields of UnauthenticatedClientPubkey

0: ShortHexStr1: String

UnauthenticatedClient(ShortHexStr, PeerId)

Tuple Fields of UnauthenticatedClient

0: ShortHexStr1: PeerId

ClientPeerIdMismatch(ShortHexStr, PeerId, PeerId)

Tuple Fields of ClientPeerIdMismatch

0: ShortHexStr1: PeerId2: PeerId

MissingAntiReplayTimestamp(ShortHexStr)

Tuple Fields of MissingAntiReplayTimestamp

0: ShortHexStr

ServerReplayDetected(ShortHexStr, u64)

Tuple Fields of ServerReplayDetected

0: ShortHexStr1: u64

BuildServerHandshakeMessageFailed(ShortHexStr, NoiseError)

Tuple Fields of BuildServerHandshakeMessageFailed

0: ShortHexStr1: NoiseError

ServerWriteFailed(ShortHexStr, Error)

Tuple Fields of ServerWriteFailed

0: ShortHexStr1: Error

Implementations

impl NoiseHandshakeError

pub fn should_security_log(&self) -> bool

Errors that are either clear bugs or indicate some security issue. Should immediately alert an engineer if we hit one of these errors.