Module network::peer_manager

The PeerManager module is responsible for establishing connections between Peers and for opening/receiving new substreams on those connections.

Implementation

The PeerManager is implemented as a number of actors:

A main event loop actor which is responsible for handling requests and sending notification about new/lost Peers to the rest of the network stack.

An actor responsible for dialing and listening for new connections.

Modules

builder

conn_notifs_channel

conn_notifs_channel is a channel which delivers to the receiver only the last of N messages that might have been sent by sender(s) since the last poll. The items are separated using a key that is provided by the sender with each message.

Structs

ConnectionRequestSender

Convenience wrapper which makes it easy to issue connection requests and await the responses from PeerManager.

PeerManager

Responsible for handling and maintaining connections to other Peers

PeerManagerRequestSender

Convenience wrapper which makes it easy to issue communication requests and await the responses from PeerManager.

Enums

ConnectionNotification

ConnectionRequest

PeerManagerError

PeerManagerNotification

Notifications sent by PeerManager to upstream actors.

PeerManagerRequest

Request received by PeerManager from upstream actors.

TransportNotification

Type Definitions

IpAddrTokenBucketLimiter