Module network::protocols::direct_send

Protocol for fire-and-forget style message delivery to a peer

DirectSend protocol takes advantage of muxers and substream negotiation to build a simple best effort message delivery protocol. Concretely,

Every message runs in its own ephemeral substream. The substream is directional in the way that only the dialer sends a message to the listener, but no messages or acknowledgements sending back on the other direction. So the message delivery is best effort and not guaranteed. Because the substreams are independent, there is no guarantee on the ordering of the message delivery either.

An DirectSend call negotiates which protocol to speak using protocol-select. This allows simple versioning of message delivery and negotiation of which message types are supported. In the future, we can potentially support multiple backwards-incompatible versions of any messages.

The actual structure of the wire messages is left for higher layers to specify. The DirectSend protocol is only concerned with shipping around opaque blobs. Current libra DirectSend clients (consensus, mempool) mostly send protobuf enums around over a single DirectSend protocol, e.g., /libra/consensus/direct_send/0.1.0.

Wire Protocol (dialer):

To send a message to a remote peer, the dialer

Requests a new outbound substream from the muxer.

Negotiates the substream using protocol-select to the protocol they wish to speak, e.g., /libra/mempool/direct_send/0.1.0.

Sends the serialized message on the newly negotiated substream.

Drops the substream.

Wire Protocol (listener):

To receive a message from remote peers, the listener

Polls for new inbound substreams on the muxer.

Negotiates inbound substreams using protocol-select. The negotiation must only succeed if the requested protocol is actually supported.

Awaits the serialized message on the newly negotiated substream.

Drops the substream.

Structs

DirectSend
The DirectSend actor.

Message

Enums

DirectSendNotification

DirectSendRequest