Module network::protocols::discovery

Discovery

The discovery module is responsible for discovering the network addresses of all members of the Libra network. Additionally, it passes on this infromation to the ConnectivityManager module to ensure appropriate connectivity.

Implementation

The discovery module is implemented as a stand-alone actor in the Network sub-system of the Libra stack. The actor participates in discovery by periodically sending its observed state of the network to a randomly chosen peer. Other peers are also expected to be running the same protocol. Therefore, in expectation, every peer expects to hear from 1 other peer in each round. On hearing from the remote peer, the local discovery module tries to reconcile its state to reflect any changes. In addition to updating its state, it also passes on new infromation to the ConnectivityManager module.

For the initial bootstrap of a node, it sends the discovery message to a randomly chosen seed peer in each round. The message only contains the identity of this peer unless it learns more about the network membership from another peer.

Currently we do not use this mechanism to detect peer failures - instead, we simply connect to all the peers in the network, and hope to learn about their failure on connection errors.

TODO: We need to handle to case of peers who may no longer be a part of the network.

Future work

Currently, we do not try to detect/punish nodes which are just lurking (without contributing to the protocol), or actively trying to spread misinformation in the network. In the future, we plan to remedy this by introducing a module dedicated to detecting byzantine behavior, and by making the discovery protocol itself tolerant to byzantine faults.

As an optimization, instead of creating a new substream to the chosen peer in each round, we could maintain a cache of open substreams which could be re-used across numerous rounds.

Structs

Discovery
Represents a prototype for an upgrade to handle the Libra Discovery protocol.

Constants

DISCOVERY_PROTOCOL_NAME