Struct network::protocols::network::NetworkEvents

pub struct NetworkEvents<TMessage> { /* fields omitted */ }

A Stream of Event<TMessage> from the lower network layer to an upper network application that deserializes inbound network direct-send and rpc messages into TMessage. Inbound messages that fail to deserialize are logged and dropped.

NetworkEvents is really just a thin wrapper around a channel::Receiver<PeerNotification> that deserializes inbound messages.