Module network::protocols::rpc

Protocol for making and handling Remote Procedure Calls

SLURP: Simple Libra Unary Rpc Protocol

SLURP takes advantage of muxers and substream negotiation to build a simple rpc protocol. Concretely,

Every rpc call runs in its own substream. Instead of managing a completion queue of message ids, we instead delegate this handling to the muxer so that the underlying substream controls the lifetime of the rpc call. Additionally, on certain transports (e.g., QUIC) we avoid head-of-line blocking as the substreams are independent.

An rpc method call negotiates which method to call using protocol-select. This allows simple versioning of rpc methods and negotiation of which methods are supported. In the future, we can potentially support multiple backwards-incompatible versions of any rpc method.

The actual structure of the request/response wire messages is left for higher layers to specify. The rpc protocol is only concerned with shipping around opaque blobs. Current libra rpc clients (consensus, mempool) mostly send protobuf enums around over a single rpc protocol, e.g., /libra/consensus/rpc/0.1.0.

Wire Protocol (dialer):

To make an rpc request to a remote peer, the dialer

Requests a new outbound substream from the muxer.

Negotiates the substream using protocol-select to the rpc method they wish to call, e.g., /libra/mempool/rpc/0.1.0.

Sends the serialized request arguments on the newly negotiated substream.

Half-closes their output side.

Awaits the serialized response message from remote.

Awaits the listener's half-close to complete the substream close.

Wire Protocol (listener):

To handle new rpc requests from remote peers, the listener

Polls for new inbound substreams on the muxer.

Negotiates inbound substreams using protocol-select. The negotiation must only succeed if the requested rpc method is actually supported.

Awaits the serialized request arguments on the newly negotiated substream.

Awaits the dialer's half-close.

Handles the request by sending it up through the NetworkProvider actor to a higher layer rpc client like consensus or mempool, who then sends the serialed rpc response back down to the rpc layer.

Sends the serialized response message to the dialer.

Half-closes their output side to complete the substream close.

Modules

error
Rpc protocol errors

fuzzing
fuzzing module for the rpc protocol

utils

Structs

InboundRpcRequest
A wrapper struct for an inbound rpc request and its associated context.

OutboundRpcRequest
A wrapper struct for an outbound rpc request and its associated context.

Rpc
The rpc actor.

Enums

RpcNotification
Events sent from the [Rpc] actor to the NetworkProvider actor.

RpcRequest
Events sent from the NetworkProvider actor to the [Rpc] actor.