Module network::protocols::wire::messaging::v1

This module defines the DiemNet v1 message types, how they are serialized/deserialized, and provides a Sink and Stream implementation for sending NetworkMessages over an abstract IO object (presumably a socket).

The DiemNet specification describes in greater detail how these messages are sent and received over-the-wire.

Structs

DirectSendMsg

NetworkMessageSink

A Sink of outbound NetworkMessages that will be serialized and sent over an underlying socket.

NetworkMessageStream

A Stream of inbound NetworkMessages read and deserialized from an underlying socket.

ParsingErrorType

Flags an invalid network message with as much header information as possible. This is a message that this peer cannot even parse its header information.

RpcRequest

RpcResponse

Enums

ErrorCode

NetworkMessage

Most primitive message type set on the network.

NotSupportedType

Flags an unsupported network message. This is a message that a peer can parse its header information but does not have a handler.

ReadError

Errors from reading and deserializing network messages off the wire.

WriteError

Errors from serializing and sending network messages on the wire.

Functions

network_message_frame_codec

Returns a fully configured length-delimited codec for writing/reading serialized NetworkMessage frames to/from a socket.

Type Definitions

Priority

Create alias Priority for u8.

RequestId

Create alias RequestId for u32.