network

Module service

Structs

Datagram

HostMetadata

NetworkContext

Wrapper around network service.

NetworkService

NetworkService implements the P2P communication between different nodes. It manages connections between peers, including accepting new peers or dropping existing peers. Inside NetworkService, it has an IoService event loop with a thread pool.

NetworkServiceInner

The inner implementation of NetworkService. Note that all accesses to the RWLocks of the fields have to follow the defined order to avoid race

ProtocolVersion

UdpChannel

UdpIoContext

Constants

DEFAULT_CONNECTION_LIFETIME_FOR_PROMOTION

DEFAULT_DISCOVERY_REFRESH_TIMEOUT

DEFAULT_DISCOVERY_ROUND_TIMEOUT

DEFAULT_FAST_DISCOVERY_REFRESH_TIMEOUT

DEFAULT_HOUSEKEEPING_TIMEOUT

DEFAULT_NODE_TABLE_TIMEOUT

MAX_DATAGRAM_SIZE

UDP_PROTOCOL_DISCOVERY

Functions

load_pos_private_key