Module network::testutils::fake_socket

This module exposes two types of sockets useful for tests:

ReadOnlyTestSocket: a socket that can be read from in different ways.

ReadWriteTestSocket: a similar wrapper but around MemorySocket to retrieve handshake messages being sent as well.

Structs

ReadOnlyTestSocket

ReadWriteTestSocket