Trait network::UnreliableNetwork

pub trait UnreliableNetwork: BaseNetwork {
    fn unreliable_send_message<'life0, 'async_trait>(
        &'life0 mut self,
        address: Multiaddr,
        message: BincodeEncodedPayload
    ) -> Pin<Box<dyn Future<Output = JoinHandle<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn unreliable_send<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        address: Multiaddr,
        message: &'life1 Self::Message
    ) -> Pin<Box<dyn Future<Output = JoinHandle<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn unreliable_broadcast<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        addresses: Vec<Multiaddr>,
        message: &'life1 Self::Message
    ) -> Pin<Box<dyn Future<Output = Vec<JoinHandle<()>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

Required Methods

fn unreliable_send_message<'life0, 'async_trait>(
&'life0 mut self,
address: Multiaddr,
message: BincodeEncodedPayload
) -> Pin<Box<dyn Future<Output = JoinHandle<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Provided Methods

fn unreliable_send<'life0, 'life1, 'async_trait>(
&'life0 mut self,
address: Multiaddr,
message: &'life1 Self::Message
) -> Pin<Box<dyn Future<Output = JoinHandle<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

fn unreliable_broadcast<'life0, 'life1, 'async_trait>(
&'life0 mut self,
addresses: Vec<Multiaddr>,
message: &'life1 Self::Message
) -> Pin<Box<dyn Future<Output = Vec<JoinHandle<()>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

Broadcasts a message to all addresses passed as an argument. The attempts to send individual messages are best effort and will not be retried.