Module network::transport

Structs

Connection

The Connection struct consists of connection metadata and the actual socket for communication.

ConnectionId

Unique local identifier for a connection.

ConnectionMetadata

Metadata associated with an established and fully upgraded connection.

DiemNetTransport

The common DiemNet Transport.

UpgradeContext

Common context for performing both inbound and outbound connection upgrades.

Constants

DIEM_TCP_TRANSPORT

tcp::Transport with Diem-specific configuration applied.

SUPPORTED_MESSAGING_PROTOCOL

Currently supported messaging protocol version. TODO: Add ability to support more than one messaging protocol.

TRANSPORT_TIMEOUT

A timeout for the connection to open and complete all of the upgrade steps.

Traits

TSocket

A trait alias for “socket-like” things.

Functions

upgrade_outbound

Upgrade an inbound connection. This means we run a Noise IK handshake for authentication and then negotiate common supported protocols.