Module network::validator_network::network_builder

Remotely authenticated vs. unauthenticated network end-points:

A network end-point operates with remote authentication if it only accepts connections from a known set of peers (trusted_peers) identified by their network identity keys. This does not mean that the other end-point of a connection also needs to operate with authentication -- a network end-point running with remote authentication enabled will connect to or accept connections from an end-point running in authenticated mode as long as the latter is in its trusted peers set.

Structs

NetworkBuilder
Build Network module with custom configuration values. Methods can be chained in order to set the configuration values. MempoolNetworkHandler and ConsensusNetworkHandler are constructed by calling NetworkBuilder::build. New instances of NetworkBuilder are obtained via NetworkBuilder::new.

Enums

TransportType
The type of the transport layer, i.e., running on memory or TCP stream, with or without Noise encryption

Constants

CONNECTIVITY_CHECK_INTERNAL_MS

DISCOVERY_INTERVAL_MS

DISOVERY_MSG_TIMEOUT_MS

INBOUND_RPC_TIMEOUT_MS

MAX_CONCURRENT_INBOUND_RPCS

MAX_CONCURRENT_NETWORK_NOTIFS

MAX_CONCURRENT_NETWORK_REQS

MAX_CONCURRENT_OUTBOUND_RPCS

MAX_CONNECTION_DELAY_MS

NETWORK_CHANNEL_SIZE

PING_FAILURES_TOLERATED

PING_INTERVAL_MS

PING_TIMEOUT_MS