Struct network::validator_network::network_builder::NetworkBuilder

pub struct NetworkBuilder { /* fields omitted */ }

Build Network module with custom configuration values. Methods can be chained in order to set the configuration values. MempoolNetworkHandler and ConsensusNetworkHandler are constructed by calling NetworkBuilder::build. New instances of NetworkBuilder are obtained via NetworkBuilder::new.

Methods

impl NetworkBuilder

pub fn new(
executor: TaskExecutor,
peer_id: PeerId,
addr: Multiaddr
) -> NetworkBuilder

Return a new NetworkBuilder initialized with default configuration values.

pub fn transport(&mut self, transport: TransportType) -> &mut Self

Set transport type, i.e., Memory or Tcp transports.

pub fn advertised_address(&mut self, advertised_address: Multiaddr) -> &mut Self

Set and address to advertise, if different from the listen address

pub fn trusted_peers(
&mut self,
trusted_peers: HashMap<PeerId, NetworkPublicKeys>
) -> &mut Self

Set trusted peers.

pub fn signing_keys(&mut self, keys: (PrivateKey, PublicKey)) -> &mut Self

Set signing keys of local node.

pub fn identity_keys(&mut self, keys: (PrivateKey, PublicKey)) -> &mut Self

Set identity keys of local node.

pub fn seed_peers(
&mut self,
seed_peers: HashMap<PeerId, Vec<Multiaddr>>
) -> &mut Self

Set seed peers to bootstrap discovery

pub fn discovery_interval_ms(&mut self, discovery_interval_ms: u64) -> &mut Self

Set discovery ticker interval

pub fn ping_interval_ms(&mut self, ping_interval_ms: u64) -> &mut Self

Set ping interval.

pub fn ping_timeout_ms(&mut self, ping_timeout_ms: u64) -> &mut Self

Set ping timeout interval.

pub fn connectivity_check_interval_ms(
&mut self,
connectivity_check_interval_ms: u64
) -> &mut Self

Set connectivity check ticker interval

pub fn inbound_rpc_timeout_ms(
&mut self,
inbound_rpc_timeout_ms: u64
) -> &mut Self

Set inbound rpc timeout.

pub fn channel_size(&mut self, channel_size: usize) -> &mut Self

Set the size of the channels between different network actors.

pub fn mempool_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that Mempool subscribes.

pub fn consensus_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that Consensus subscribes.

pub fn state_synchronizer_protocols(
&mut self,
protocols: Vec<ProtocolId>
) -> &mut Self

Set the protocol IDs that StateSynchronizer subscribes.

pub fn direct_send_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that DirectSend actor subscribes.

pub fn rpc_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that RPC actor subscribes.

pub fn build(
&mut self
) -> ((MempoolNetworkSender, MempoolNetworkEvents), (ConsensusNetworkSender, ConsensusNetworkEvents), (StateSynchronizerNetworkSender, StateSynchronizerNetworkEvents), Multiaddr)

Create the configured NetworkBuilder Return the constructed Mempool and Consensus Sender+Events