Crate network

Modules

connectivity_manager
The ConnectivityManager actor is responsible for ensuring that we are connected to a node if and only if it is an eligible node.

constants

counters

error

fuzzing

logging
This module is to contain all networking logging information.

noise
This crate implements wrappers around our Noise implementation. Noise is a protocol framework to encrypt and authentication connections. We use Noise to secure connections between peers in Diem. Specifically, we use the Noise IK handshake which is a one round-trip protocol (the client sends one message, then the server responds). For more information about Noise and our implementation, refer to the crypto crate.

peer
Peer manages a single connection to a remote peer after the initial connection establishment and handshake.

peer_manager
The PeerManager module is responsible for establishing connections between Peers and for opening/receiving new substreams on those connections.

protocols
Protocols used by network module for external APIs and internal functionality

testutils

transport

Type Definitions

ConnectivityRequest

DisconnectReason

ProtocolId