Crate network

Re-exports

pub use common::NetworkPublicKeys;

pub use interface::NetworkProvider;

Modules

common

connectivity_manager
The ConnectivityManager actor is responsible for ensuring that we are connected to a node if and only if it is an eligible node.

error

interface
Module exposing generic network interface to a single connected peer.

peer_manager
The PeerManager module is responsible for establishing connections between Peers and for opening/receiving new substreams on those connections.

protocols
Protocols used by network module for external APIs and internal functionality

validator_network
Network API for Libra

Type Definitions

ConnectivityRequest

DisconnectReason

ProtocolId