Crate network

Re-exports

pub use validator_network::ValidatorNetwork;

Modules

interface
Internal Rust Network API for Consensus and Mempool

multiplexing
The Multiplexing module is responsible for defining and implementing Stream Multiplexing.

negotiate
Everything to do with negotiating protocols on AsyncRead/AsyncWrite streams

noise
Noise protocol framework support for use in Libra.

proto

protocols

transport
The Transport module is responsible for establishing connections with Peers.

validator_network
Validator Network

Structs

NetworkPublicKeys
Public keys used at the network layer.

Type Definitions

ProtocolId