Crate network

Modules

metrics

Structs

BoundedExecutor

CancelOnDropHandler

This adapter will make a [tokio::task::JoinHandle] abort its handled task when the handle is dropped.

PrimaryNetwork

PrimaryToWorkerNetwork

RetryConfig

Retry configurations for establishing connections and sending messages. Determines the retry behaviour of requests, by setting the back off strategy used.

WorkerNetwork

WorkerToPrimaryNetwork

Constants

MAX_TASK_CONCURRENCY

Traits

Lucky

LuckyNetwork

LuckyNetwork2

ReliableNetwork

ReliableNetwork2

UnreliableNetwork

UnreliableNetwork2

Functions

multiaddr_to_address

Type Definitions

MessageResult