Module network::counters

Constants

CANCELED_LABEL

DECLINED_LABEL

FAILED_LABEL

RECEIVED_LABEL

REQUEST_LABEL

RESPONSE_LABEL

SENT_LABEL

Statics

LIBRA_NETWORK_DIRECT_SEND_BYTES

LIBRA_NETWORK_DIRECT_SEND_MESSAGES

LIBRA_NETWORK_DISCOVERY_NOTES

LIBRA_NETWORK_PEERS

LIBRA_NETWORK_RPC_BYTES

LIBRA_NETWORK_RPC_LATENCY

LIBRA_NETWORK_RPC_MESSAGES

OP_COUNTERS

PENDING_CONNECTION_HANDLER_NOTIFICATIONS
Counter of pending Connection Handler notifications to PeerManager.

PENDING_CONNECTIVITY_MANAGER_REQUESTS
Channel Counters

PENDING_DIRECT_SEND_NOTIFICATIONS
Counter of pending Direct Send notifications to Network Provider

PENDING_DIRECT_SEND_REQUESTS
Counter of pending requests in Direct Send

PENDING_DISCOVERY_NETWORK_EVENTS
Counter of pending network events to Discovery.

PENDING_HEALTH_CHECKER_NETWORK_EVENTS
Counter of pending network events to Health Checker.

PENDING_NETWORK_NOTIFICATIONS
Counters(queued,dequeued,dropped) related to inbound network notifications for RPCs and DirectSends.

PENDING_NETWORK_REQUESTS
Counter of pending requests in Network Provider

PENDING_PEER_DIRECT_SEND_NOTIFICATIONS
Counter of pending DirectSend events from Peer to DirectSend actor..

PENDING_PEER_MANAGER_DIAL_REQUESTS
Counter of pending dial requests in Peer Manager

PENDING_PEER_MANAGER_REQUESTS
Counter of pending requests in Peer Manager

PENDING_PEER_NETWORK_NOTIFICATIONS
Counter of pending connection notifications from Peer to NetworkProvider.

PENDING_PEER_REQUESTS
Counter of pending requests for each remote peer

PENDING_PEER_RPC_NOTIFICATIONS
Counter of pending RPC events from Peer to Rpc actor.

PENDING_RPC_NOTIFICATIONS
Counter of pending RPC notifications to Network Provider

PENDING_RPC_REQUESTS
Counter of pending requests in RPC

PENDING_WIRE_MESSAGES
Counter of messages pending in queue to be sent out on the wire.