Module network::interface

Module exposing generic network interface to a single connected peer.

Unlike the validator_network module, which exposes async function call and Stream API specific for consensus and mempool modules, the interface module exposes generic network API by receiving requests over a channel for outbound requests and sending notifications to upstream clients for inbound requests and other events. For example, clients wishing to send an RPC need to send a NetworkRequest::SendRpc message to the NetworkProvider actor. Inbound RPC requests are forwarded to the appropriate handler, determined using the protocol negotiated on the RPC substream.

Structs

NetworkProvider

Enums

NetworkNotification
Notifications that NetworkProvider sends to consumers of its API. The NetworkProvider in turn receives these notifications from the PeerManager and other protocols.

NetworkRequest
Requests NetworkProvider receives from the network interface.