Module network::logging

This module is to contain all networking logging information.

use libra_config::network_id::NetworkContext;
use libra_logger::prelude::*;
use network::logging::*;

send_struct_log!(
  network_log(network_events::CONNECTIVITY_MANAGER_LOOP, &NetworkContext::mock())
    .data(network_events::TYPE, network_events::START)
    .field(network_events::EVENT_ID, &5)
);

Modules

network_events
This module is to ensure no conflicts with already existing constants

Functions

network_log
A helper function to cut down on a bunch of repeated network struct log code