Module network::noise::handshake

The handshake module implements the handshake part of the protocol. This module also implements additional anti-DoS mitigation, by including a timestamp in each handshake initialization message. Refer to the module’s documentation for more information. A successful handshake returns a NoiseStream which is defined in the stream module.

Structs

AntiReplayTimestamps

In a mutually authenticated network, a client message is accompanied with a timestamp. This is in order to prevent replay attacks, where the attacker does not know the client’s static key, but can still replay a handshake message in order to force a peer into performing a few Diffie-Hellman key exchange operations.

NoiseUpgrader

The Noise configuration to be used to perform a protocol upgrade on an underlying socket.

Enums

HandshakeAuthMode

Noise handshake authentication mode.