Module network::protocols::direct_send

Implementation of DirectSend as per Libra wire protocol v1.

Design:

DirectSend exposes a very simple API to clients. It receives SendMessage requests from upstream clients for outbound message while sending RecvMessage notifications on receipt of inbound messages. The DirectSend actor runs in a single event loop where it handles notifications about new inbound messages from the Peer actor and requests for outbound messages from upstream clients. Both types of messages are processed inline instead of being spawned into separate tasks, since they both simply entail forwarding of messages.

Limits:

Since DirectSend does not dedicate any resources to inbound messages (except forwarding them to the upstream client), it does not try to limit them in any way. Rate-limiting of overall number of messages received at the Peer actor should be sufficient for safe-guarding against malicious actors.

Structs

DirectSend
The DirectSend actor.

Message

Enums

DirectSendNotification

DirectSendRequest