Module network::protocols::discovery

Protocol to discover network addresses of other peers on the Libra network

Implementation

The discovery module is implemented as a stand-alone actor in the Network sub-system of the Libra stack. The actor participates in discovery by periodically sending its observed state of the network to a randomly chosen peer. Other peers are also expected to be running the same protocol. Therefore, in expectation, every peer expects to hear from 1 other peer in each round. On hearing from the remote peer, the local discovery module tries to reconcile its state to reflect any changes. In addition to updating its state, it also passes on new information to the ConnectivityManager module.

For the initial bootstrap of a node, it sends the discovery message to a randomly chosen seed peer in each round. The message only contains the identity of this peer unless it learns more about the network membership from another peer.

Currently we do not use this mechanism to detect peer failures - instead, we simply connect to all the peers in the network, and hope to learn about their failure on connection errors.

TODO: We need to handle to case of peers who may no longer be a part of the network.

Future work

Currently, we do not try to detect/punish nodes which are just lurking (without contributing to the protocol), or actively trying to spread misinformation in the network. In the future, we plan to remedy this by introducing a module dedicated to detecting byzantine behavior, and by making the discovery protocol itself tolerant to byzantine faults.

Structs

Discovery
The actor running the discovery protocol.

DiscoveryMsg
A Discovery message contains notes collected from other peers within the system.

DiscoveryNetworkSender
The interface from Discovery to Networking layer.

FullNodeInfo
Discovery information relevant to public full nodes and clients.

FullNodeInfoHasher

Note
A Note contains a validator's signed PeerInfo as well as a signed FullNodePayload, which provides relevant discovery info for public full nodes and clients.

PeerInfo
A PeerInfo represents the network address(es) of a Peer.

PeerInfoHasher

SignedFullNodeInfo
A FullNodeInfo authenticated by the peer's root network_signing_key stored on-chain.

SignedPeerInfo
A PeerInfo authenticated by the peer's root network_signing_key stored on-chain.

VerifiedNote

Functions

add_to_network
Register the discovery sender and event handler with network and return interfaces for those actors.

Type Definitions

DiscoveryNetworkEvents
The interface from Network to Discovery module.