Module network::protocols::network

Convenience Network API for Diem

Re-exports

pub use crate::protocols::rpc::error::RpcError;

Structs

AppConfig

Configuration needed for DiemNet applications to register with the network builder. Supports client-only, service-only, and p2p (both) applications.

NetworkEvents

A Stream of Event<TMessage> from the lower network layer to an upper network application that deserializes inbound network direct-send and rpc messages into TMessage. Inbound messages that fail to deserialize are logged and dropped.

NetworkSender

NetworkSender is the generic interface from upper network applications to the lower network layer. It provides the full API for network applications, including sending direct-send messages, sending rpc requests, as well as dialing or disconnecting from peers and updating the list of accepted public keys.

Enums

Event

Events received by network clients in a validator

Traits

ApplicationNetworkSender

A simplified version of NetworkSender that doesn’t use ProtocolId in the input It was already being implemented for every application, but is now standardized

Message

NewNetworkEvents

Trait specifying the signature for new() NetworkEvents

NewNetworkSender

Trait specifying the signature for new() NetworkSenders

SerializedRequest

Generalized functionality for any request across DirectSend and Rpc.