Module network::protocols::network

Convenience Network API for Libra

Re-exports

pub use crate::protocols::rpc::error::RpcError;

Structs

NetworkEvents
A Stream of Event<TMessage> from the lower network layer to an upper network application that deserializes inbound network direct-send and rpc messages into TMessage, which is some protobuf format implementing the prost::Message trait.

NetworkSender
NetworkSender is the generic interface from upper network applications to the lower network layer. It provides the full API for network applications, including sending direct-send messages, sending rpc requests, as well as dialing or disconnecting from peers and updating the list of accepted public keys.

Enums

Event
Events received by network clients in a validator

Traits

Message