Module network::protocols::rpc

Implementation of the RPC protocol as per Libra wire protocol v1.

Design:

RPC receives OutboundRpcRequest messages from upstream actors. The OutboundRpcRequest contains the RPC protocol, raw request bytes, RPC timeout duration, and a channel over which the response bytes can be sent back to the upstream. For inbound RPC requests, RPC sends InboundRpcRequest notifications to upstream actors. The InboundRpcRequest contains the RPC protocol, raw request bytes, and a channel over which the upstream can send a response for the RPC. Internally, the RPC actor consists of a single event loop. The event loop processes 4 kinds of messages: (1) outbound RPC requests received from upstream, (2) notifications for inbound RpcRequest/RpcResponse from the Peer actor, (3) completion notification for tasks processing inbound RPC, and (4) completion notification for tasks processing outbound RPCs. The tasks for inbound and outbound RPCs are spawned onto the same runtime as the one driving the RPC event loop.

Timeouts:

The tasks for inbound and outbound RPCs are also "wrapped" within timeouts to ensure that they are not running forever. The outbound RPC timeout is specified by the upstream client, where as the inbound RPC timeout is a configuration parameter for the RPC actor.

Limits:

We limit the number of pending inbound RPC tasks to ensure that resource usage is bounded for inbound RPCs. For outbound RPCs, we log a warning when the limit is exceeded, but allow the RPC to proceed.

State

For outbound RPCs, the RPC actors maintains a HashMap from the RequestId to a channel over which inbound responses can be delivered to the task driving the request. Entries are removed on completion of the task, which happens either on receipt of the response, or on failure/timeout.

The RPC actor also maintains a RequestIdGenerator for generating request ids for outbound RPCs. The RequestIdGenerator increments the request id by 1 for each subsequent outbound RPC.

Modules

error
Rpc protocol errors

fuzzing
fuzzing module for the rpc protocol

Structs

InboundRpcRequest
A wrapper struct for an inbound rpc request and its associated context.

OutboundRpcRequest
A wrapper struct for an outbound rpc request and its associated context.

Rpc
The rpc actor.

Enums

RpcNotification
Events sent from the Rpc actor to the NetworkProvider actor.