Module network::protocols::wire::messaging::v1

This module defines the structs transported during the network messaging protocol v1. These should serialize as per [link](TODO: Add ref).

Structs

DirectSendMsg

Nonce
Nonces used by Ping and Pong message types.

RpcRequest

RpcResponse

Enums

ErrorCode
Enum representing various error codes that can be embedded in NetworkMessage. New variants cannot be added without bumping up the MessagingProtocolVersion.

NetworkMessage
Message variants that are sent on the wire. New variants cannot be added without bumping up the MessagingProtocolVersion.

Type Definitions

Priority
Create alias Priority for u8.

RequestId
Create alias RequestId for u32.