Module network::transport

Structs

Connection
The Connection struct consists of connection metadata and the actual socket for communication.

ConnectionId
Unique local identifier for a connection.

ConnectionMetadata
Metadata associated with an established and fully upgraded connection.

LibraNetTransport
The common LibraNet Transport.

Constants

LIBRA_TCP_TRANSPORT
tcp::Transport with Libra-specific configuration applied.

SUPPORTED_MESSAGING_PROTOCOL
Currently supported messaging protocol version. TODO: Add ability to support more than one messaging protocol.

TRANSPORT_TIMEOUT
A timeout for the connection to open and complete all of the upgrade steps.

Traits

TSocket
A trait alias for "socket-like" things.

Functions

perform_handshake
Exchange HandshakeMsg's to try negotiating a set of common supported protocols.