Struct network::validator_network::network_builder::NetworkBuilder

pub struct NetworkBuilder { /* fields omitted */ }

Build Network module with custom configuration values. Methods can be chained in order to set the configuration values. MempoolNetworkHandler and ConsensusNetworkHandler are constructed by calling NetworkBuilder::build. New instances of NetworkBuilder are obtained via NetworkBuilder::new.

Methods

impl NetworkBuilder

pub fn new(
executor: Handle,
peer_id: PeerId,
addr: NetworkAddress,
role: RoleType
) -> NetworkBuilder

Return a new NetworkBuilder initialized with default configuration values.

pub fn transport(&mut self, transport: TransportType) -> &mut Self

Set transport type, i.e., Memory or Tcp transports.

pub fn advertised_address(
&mut self,
advertised_address: NetworkAddress
) -> &mut Self

Set and address to advertise, if different from the listen address

pub fn trusted_peers(
&mut self,
trusted_peers: HashMap<PeerId, NetworkPublicKeys>
) -> &mut Self

Set trusted peers.

pub fn signing_keypair(
&mut self,
keypair: (Ed25519PrivateKey, Ed25519PublicKey)
) -> &mut Self

Set signing keys of local node.

pub fn seed_peers(
&mut self,
seed_peers: HashMap<PeerId, Vec<NetworkAddress>>
) -> &mut Self

Set seed peers to bootstrap discovery

pub fn discovery_interval_ms(&mut self, discovery_interval_ms: u64) -> &mut Self

Set discovery ticker interval

pub fn ping_interval_ms(&mut self, ping_interval_ms: u64) -> &mut Self

Set ping interval.

pub fn ping_failures_tolerated(
&mut self,
ping_failures_tolerated: u64
) -> &mut Self

Set number of ping failures tolerated.

pub fn ping_timeout_ms(&mut self, ping_timeout_ms: u64) -> &mut Self

Set ping timeout.

pub fn connectivity_check_interval_ms(
&mut self,
connectivity_check_interval_ms: u64
) -> &mut Self

Set connectivity check ticker interval

pub fn inbound_rpc_timeout_ms(
&mut self,
inbound_rpc_timeout_ms: u64
) -> &mut Self

Set inbound rpc timeout.

pub fn max_concurrent_outbound_rpcs(
&mut self,
max_concurrent_outbound_rpcs: u32
) -> &mut Self

The maximum number of concurrent outbound rpc requests we will service.

pub fn max_concurrent_inbound_rpcs(
&mut self,
max_concurrent_inbound_rpcs: u32
) -> &mut Self

The maximum number of concurrent inbound rpc requests we will service.

pub fn max_concurrent_network_reqs(
&mut self,
max_concurrent_network_reqs: usize
) -> &mut Self

The maximum number of concurrent NetworkRequests we will service in NetworkProvider.

pub fn max_concurrent_network_notifs(
&mut self,
max_concurrent_network_notifs: usize
) -> &mut Self

The maximum number of concurrent Notifications from each actor we will service in NetworkProvider.

pub fn max_connection_delay_ms(
&mut self,
max_connection_delay_ms: u64
) -> &mut Self

The maximum duration (in milliseconds) we should wait before dialing a peer we should connect to.

pub fn channel_size(&mut self, channel_size: usize) -> &mut Self

Set the size of the channels between different network actors.

pub fn enable_remote_authentication(
&mut self,
enable_remote_authentication: bool
) -> &mut Self

Set the enable_remote_authentication flag to make the network operate with remote authentication.

pub fn conn_mgr_reqs_tx(&self) -> Option<Sender<ConnectivityRequest>>

pub fn add_protocol_handler(
&mut self,
rpc_protocols: Vec<ProtocolId>,
direct_send_protocols: Vec<ProtocolId>,
queue_preference: QueueStyle,
max_queue_size_per_peer: usize,
counter: Option<&'static IntCounterVec>
) -> (PeerManagerRequestSender, Receiver<(PeerId, ProtocolId), PeerManagerNotification>, ConnectionRequestSender, Receiver)

Add a handler for given protocols using raw bytes.

pub fn add_connection_event_listener(&mut self) -> Receiver

pub fn add_connectivity_manager(&mut self) -> &mut Self

Add a ConnectivityManager to the network.

ConnectivityManager is responsible for ensuring that we are connected to a node iff. it is an eligible node and maintaining persistent connections with all eligible nodes. A list of eligible nodes is received at initialization, and updates are received on changes to system membership.

Note: a connectivity manager should only be added if the network is permissioned.

pub fn add_gossip_discovery(&mut self) -> &mut Self

Add the (gossip) Discovery protocol to the network.

(gossip) Discovery discovers other eligible peers' network addresses by exchanging the full set of known peer network addresses with connected peers as a network protocol.

pub fn add_connection_monitoring(&mut self) -> &mut Self

pub fn build(self) -> NetworkAddress

Create the configured transport and start PeerManager. Return the actual NetworkAddress over which this peer is listening.