Struct network::validator_network::network_builder::NetworkBuilder

pub struct NetworkBuilder { /* fields omitted */ }

Build Network module with custom configuration values. Methods can be chained in order to set the configuration values. MempoolNetworkHandler and ConsensusNetworkHandler are constructed by calling [NetworkBuilder::build]. New instances of NetworkBuilder are obtained via [NetworkBuilder::new].

Methods

impl NetworkBuilder

pub fn new(
executor: TaskExecutor,
peer_id: PeerId,
addr: Multiaddr,
role: RoleType
) -> NetworkBuilder

Return a new NetworkBuilder initialized with default configuration values.

pub fn transport(&mut self, transport: TransportType) -> &mut Self

Set transport type, i.e., Memory or Tcp transports.

pub fn advertised_address(&mut self, advertised_address: Multiaddr) -> &mut Self

Set and address to advertise, if different from the listen address

pub fn trusted_peers(
&mut self,
trusted_peers: HashMap<PeerId, NetworkPublicKeys>
) -> &mut Self

Set trusted peers.

pub fn signing_keys(
&mut self,
keys: (Ed25519PrivateKey, Ed25519PublicKey)
) -> &mut Self

Set signing keys of local node.

pub fn seed_peers(
&mut self,
seed_peers: HashMap<PeerId, Vec<Multiaddr>>
) -> &mut Self

Set seed peers to bootstrap discovery

pub fn discovery_interval_ms(&mut self, discovery_interval_ms: u64) -> &mut Self

Set discovery ticker interval

pub fn ping_interval_ms(&mut self, ping_interval_ms: u64) -> &mut Self

Set ping interval.

pub fn ping_failures_tolerated(
&mut self,
ping_failures_tolerated: u64
) -> &mut Self

Set number of ping failures tolerated.

pub fn ping_timeout_ms(&mut self, ping_timeout_ms: u64) -> &mut Self

Set ping timeout.

pub fn discovery_msg_timeout_ms(
&mut self,
discovery_msg_timeout_ms: u64
) -> &mut Self

Set discovery message timeout.

pub fn connectivity_check_interval_ms(
&mut self,
connectivity_check_interval_ms: u64
) -> &mut Self

Set connectivity check ticker interval

pub fn inbound_rpc_timeout_ms(
&mut self,
inbound_rpc_timeout_ms: u64
) -> &mut Self

Set inbound rpc timeout.

pub fn max_concurrent_outbound_rpcs(
&mut self,
max_concurrent_outbound_rpcs: u32
) -> &mut Self

The maximum number of concurrent outbound rpc requests we will service.

pub fn max_concurrent_inbound_rpcs(
&mut self,
max_concurrent_inbound_rpcs: u32
) -> &mut Self

The maximum number of concurrent inbound rpc requests we will service.

pub fn max_concurrent_network_reqs(
&mut self,
max_concurrent_network_reqs: u32
) -> &mut Self

The maximum number of concurrent NetworkRequests we will service in NetworkProvider.

pub fn max_concurrent_network_notifs(
&mut self,
max_concurrent_network_notifs: u32
) -> &mut Self

The maximum number of concurrent Notifications from each actor we will service in NetworkProvider.

pub fn max_connection_delay_ms(
&mut self,
max_connection_delay_ms: u64
) -> &mut Self

The maximum duration (in milliseconds) we should wait before dialing a peer we should connect to.

pub fn channel_size(&mut self, channel_size: usize) -> &mut Self

Set the size of the channels between different network actors.

pub fn direct_send_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that DirectSend actor subscribes.

pub fn rpc_protocols(&mut self, protocols: Vec<ProtocolId>) -> &mut Self

Set the protocol IDs that RPC actor subscribes.

pub fn permissioned(&mut self, is_permissioned: bool) -> &mut Self

Set the is_permissioned flag to make the network permissioned or permission-less.

pub fn build(&mut self) -> (Multiaddr, Box<dyn LibraNetworkProvider>)

Create the configured NetworkBuilder Return the constructed Mempool and Consensus Sender+Events